Force collect global adaptivity micro data to be float32

+++ micro_manager/adaptivity/global_adaptivity.py
@@ -115,7 +115,9 @@ class GlobalAdaptivityCalculator(AdaptivityCalculator):
         global_data_for_adaptivity = dict()
         for name in data_for_adaptivity.keys():
             data_as_list = self._comm.allgather(data_for_adaptivity[name])
-            global_data_for_adaptivity[name] = np.concatenate((data_as_list[:]), axis=0)
+            global_data_for_adaptivity[name] = np.concatenate(
+                (data_as_list[:]), axis=0, dtype=np.float32
+            )
 
         self._get_similarity_dists(
             dt, self._similarity_dists, global_data_for_adaptivity
